pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flightlab-rl"
version = "0.1.0"
description = "Mac-friendly fixed-wing simulation and reinforcement learning lab."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "flightlab-rl contributors" }]
dependencies = [
  "gymnasium>=1.1.1",
  "numpy>=2.1.0",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1",
  "ruff>=0.11.5",
]
sim = [
  "jsbsim>=1.2.3",
]
rl = [
  "stable-baselines3>=2.5.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
branch = true
source = ["src/flightlab"]

[tool.coverage.report]
fail_under = 95
show_missing = true
skip_covered = false
